backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.16.1
pillow==10.2.0
numpy==1.26.4
opencv-python==4.9.0.80
groq==0.9.0
python-dotenv==1.0.1
httpx==0.27.0